week5/RAG_KNOWLEDGE_WORKER/config.py
"""Simple config for Day 5 style RAG with Groq and Python files."""

import os
from pathlib import Path
from dotenv import load_dotenv

load_dotenv()

PROJECT_ROOT = Path(__file__).parent
KNOWLEDGE_BASE_PATH = PROJECT_ROOT / "knowledge_base"
STORE_PATH = PROJECT_ROOT / "rag_store.json"

GROQ_API_KEY = os.getenv("GROQ_API_KEY", "")
GROQ_MODEL = os.getenv("GROQ_MODEL", "llama-3.3-70b-versatile")

# Retrieval settings
CHUNK_SIZE = int(os.getenv("CHUNK_SIZE", "800"))
CHUNK_OVERLAP = int(os.getenv("CHUNK_OVERLAP", "120"))
TOP_K = int(os.getenv("TOP_K", "4"))
