week5/RAG_KNOWLEDGE_WORKER/requirements.txt
groq==0.4.1
python-dotenv==1.0.0
httpx==0.27.2
